Password Generator - Şifre Oluşturucu
TÜRKÇE DİLİ
Java dili ile hazırladığım Sifre Olusturucu - Password Generator min-max (12-24) karakter olur.
şifre de özel karakterler de kullanılır kullanıcı seçenekler ile kayıt ol seçeneğini tuşlayınca gerekli bilgiler istenir şifre seçeneğinde bir daha tuşlama istenir (otomatik oluşturma, manuel oluşturma) tuşlama sonrası ve şifre istendikten sonra klasörde oluşturulan kullanicilar.txt ye bu bilgiler kayıt edilir ve son olarak bu bilgiler veritabanından çeker gibi projeyi tekrar çalıştırınca bir daha kayıt olmaya gerek kalmaz ilk belirlenen kullanıcıadı ve şifreden giriş yapılabilir. İlk Java kodum hatam olabilir kodumu incelediğinizden dolayı şükranlarımı sunarım

ENGLİSH LANGUAGE
I have prepared a Password Generator in the Java programming language, which generates passwords with a length between the minimum (12) and maximum (24) characters. When the user selects the "Sign Up" option with the choice to use special characters, the required information is requested. After selecting the password option, the user is prompted to enter the password again, with two options for automatic and manual password creation. After the user enters the password, the information is saved to a file named "kullanicilar.txt" in the folder. Lastly, when the project is run again, there is no need to sign up again. Users can log in with the initially set username and password. This might be my first Java code, so there may be some errors. I express my gratitude for reviewing my code.